# [PERF] point_of_sale, stock_account: Prefetch stock values