This commit optimizes the `_accumulate_amounts` method. Previously, when there were many stock moves per PoS order, the '_compute_average_price' would be called for each move, resulting in multiple identical queries for each valuation layer in each stock move.

Now, the stock moves and valuation layers for an order are prefetched at once, significantly reducing the number of database queries. This results in a performance improvement, especially when processing a large number of PoS orders and stock moves.

| Scenario | Before Optimization (seconds) | After Optimization (seconds) |
|----------|------------------------------|-----------------------------|
| 100 orders, 1,000 stock moves | 20 | 5 |
| 300 orders, 100,000 stock moves | 617 | 23 |

opw-3980597

---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr