@robodoo 2024-06-11
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/168963.png)](https://mergebot.odoo.com/odoo/odoo/pull/168963)