@caburj 2024-06-12
Hi @pivi-odoo. Can you get your opinion on this PR?