@caburj 2024-06-12
Hi @pedrambiria. Can you also provide some performance metrics of before and after?