@pedrambiria 2024-06-19
@pivi-odoo Thanks for your review. I modified the code.